Split version utilities into separate file
